warehouse-service/ml/ml_service_requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pandas>=2.2.2
numpy>=2.0.0
pydantic>=2.8.0
# ML stack (required by ml_service_app.py)
lightgbm>=4.3.0
joblib>=1.4.0
scikit-learn>=1.5.0
# Database access (Postgres)
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
